# Decrease allocations. Don't force caller to allocate.