@nastevens 2017-04-12
Hey @insanitybit - sorry to jump in on your PR unasked, but I had been looking at doing something almost identical so I had some ideas on this. Rather than using `&str` and slices though, I was looking at `Into<Cow<'a, str>>` as described in [From &str to Cow](http://blog.jwilm.io/from-str-to-cow/). The API when doing this seems really nice, as it pretty much makes legal all variations of `&str`, `String`, etc.

Similarly on the change from `Vec<>` to slices, I have had really good results having API functions take `I: IntoIterator<Item=S>, S: AsRef<str>`. This allows _any_ iterable of anything that can be referenced as a `str` to be used. So `Vec<String>` is still legal, but so is a static array of `&str`