@mcasper 2017-04-12
Hey @insanitybit, thanks so much for the contribution! I've been meaning to come back through and clean up allocations, this is awesome!

Don't worry about the breaking changes, we're pre 1.0 so we can do whatever we want 😛 

I do like @nastevens comments about making the API accept both forms of each argument, so if you have time to update the PR that would be great