@insanitybit 2017-04-12
I've updated the code to use Cow's. Not sure exactly how to get the API working with the IntoIterator. I'm running into some borrow errors.