@mcasper 2017-04-12
I started implementing it roughly off this branch, looks like most of the type signatures will change to be like: 
```Rust
pub fn incr<'a, S, I>(&self, stat: S, tags: I) -> DogstatsdResult
    where S: Into<Cow<'a, str>>,
               S: AsRef<str>,
               I: IntoIterator<Item=S>
    {
```

The `AsRef<str>` and `IntoIterator<Item=S>` types will basically permeate all the way down to `format_for_send`, at which point we can consume the iterator and add any tags that might be there