@mcasper 2017-04-21
Thanks again for the contribution @insanitybit! I went ahead and merged this and then pushed up the `IntoIterator` change, and published a new version with all this great work